refactoring continue

+++ Yurchenko_Final/src/main/java/controller/command/DeleteUserById.java
@@ -5,6 +5,7 @@ import model.service.impl.UserServiceImpl;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.Optional;
 
 public class DeleteUserById  extends Command implements CommandPages {
 
@@ -12,12 +13,13 @@ public class DeleteUserById  extends Command implements CommandPages {
 
     @Override
     public CommandResult execute(HttpServletRequest req, HttpServletResponse resp) {
-        Long userId = Long.parseLong(req.getParameter("radio"));
-        userServiceImpl.deleteUserById(userId);
-
+        Optional<String> optId = Optional.ofNullable(req.getParameter("radio"));
+        if(optId.isPresent()) {
+            Long userId = Long.parseLong(req.getParameter("radio"));
+            userServiceImpl.deleteUserById(userId);
+            return CommandResult.forward(new ShowAllUsers());
+        }
         return CommandResult.forward(new ShowAllUsers());
-
-
 
     }
 }
